% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{addLevel_Nottaken}
\alias{addLevel_Nottaken}
\title{prepend Not taken as extra level}
\usage{
addLevel_Nottaken(data, names)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
updated full database
}
\description{
\code{joinWithLookups} takes Access database and updates one of the tables
}

