% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{as.Date.POSIX}
\alias{as.Date.POSIX}
\title{Convert different classes to Date class.}
\usage{
\method{as.Date}{POSIX}(date)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
updated full database
}
\description{
\code{as.Date.POSIX} takes Access database and updates one of the tables
}

