% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealised-pathway-cost.R
\name{boot.calcIdealisedTestCosts}
\alias{boot.calcIdealisedTestCosts}
\title{Bootstrap calculate idealised test costs [DEPRECATED]}
\usage{
boot.calcIdealisedTestCosts(data, stat = "Mean", n = 100)
}
\arguments{
\item{data}{IDEA data set}

\item{stat}{Statistic (Mean, Median, 1st Qu., 3rd Qu.)}

\item{n}{Number of bootstrap samples}
}
\value{
Array of \code{n} bootstrap outputs
}
\description{
Bootstrap calculate idealised test costs [DEPRECATED]
}
\seealso{
\code{\link{calcIdealisedTestCosts}}.
}
