% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{boot.testPerformanceMetricTable}
\alias{boot.testPerformanceMetricTable}
\title{Make array of statistics for each bootstrap sample of IDEA study data}
\usage{
boot.testPerformanceMetricTable(n = 1000)
}
\arguments{
\item{n}{Number of bootstrap samples}
}
\value{
dim 3 array
}
\description{
Make array of statistics for each bootstrap sample of IDEA study data
}

