% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanceddiagcost_table.R
\name{bootstrap.data}
\alias{bootstrap.data}
\title{Bootstrap data}
\usage{
bootstrap.data(data)
}
\arguments{
\item{data}{}
}
\value{
Bootstrapped dataset
}
\description{
Bootstrap patient data for mean and standard deviation estimates
for current in-practice routine pathway (without rule-out test).
}

