% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanceddiagcost_table.R
\name{calc.enhancedDiagTimeAndCost.bootsample}
\alias{calc.enhancedDiagTimeAndCost.bootsample}
\title{Create an enhanced pathway individual-level dataset of time to diagnosis and test costs with random samples}
\usage{
calc.enhancedDiagTimeAndCost.bootsample(data, sens = 0.9, spec = 0.9,
  FNtime = 42, testcost = 200, prop_highrisk = 0.4)
}
\arguments{
\item{sens}{Sensitivity}

\item{spec}{Specificity}

\item{FNtime}{Follow-up time of false negatives}

\item{testcost}{Rule-out test unit cost}

\item{prop_highrisk}{Clinical judgement threshold of high risk patients}
}
\value{
Bootstrapped dataset
}
\description{
This can be used to then bootstrap patient data for mean and standard deviation estimates
for enhanced pathway (with rule-out test) by representing as mixture model.
}
