% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPI.R
\name{calc.EVPI}
\alias{calc.EVPI}
\title{Calculate Expected Value of Information}
\usage{
calc.EVPI(N = 5000)
}
\value{
array
}
\description{
Calculate Expected Value of Information
}
\seealso{
\code{\link{plot.EVPI}}
}
