% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{calc.statOutcomeByCovariate}
\alias{calc.statOutcomeByCovariate}
\title{get.statStartToDiagByCovariate}
\usage{
calc.statOutcomeByCovariate(field, value, stat, outcome)
}
\arguments{
\item{field}{}

\item{value}{}

\item{stat}{}

\item{outcome}{}
}
\value{
data
}
\description{
\code{get.statOutcomeByCovariate}
}

