% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{calc.summarystats_Dosanjh_TimeCost}
\alias{calc.summarystats_Dosanjh_TimeCost}
\title{Calculate summary statistics to include in final tables}
\usage{
calc.summarystats_Dosanjh_TimeCost(data)
}
\arguments{
\item{data}{}
}
\value{
matrix
}
\description{
\code{maketable.Dosanjh_TimeCost}
}
\seealso{

}
