% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{calc.testDate.min}
\alias{calc.testDate.min}
\title{Calculate the First Test Dates for Each Patient}
\usage{
calc.testDate.min(data, testDate.names, maxtime)
}
\arguments{
\item{data}{IDEA study data set.}

\item{testDate.names}{Names of all test dates for each patient.}

\item{maxtime}{The earliest allowable date before the date of consent for a test to be taken and still be for the same episode of TB.}
}
\value{
testDate.min
}
\description{
Which test dates are earliest within a time window from the date of study consent.
}

