% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPI.R
\name{calcExpectedINMB}
\alias{calcExpectedINMB}
\title{Calculate Expected INMB}
\usage{
calcExpectedINMB(spec = 0.9, sens = 0.9, Ctest = 200)
}
\arguments{
\item{Ctest}{}
}
\value{
INMB sample
}
\description{
Calculate Expected INMB
}
