% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealised-pathway-cost.R
\name{calcIdealisedTestCosts}
\alias{calcIdealisedTestCosts}
\title{Calculate NICE Idealised Test Costs}
\usage{
calcIdealisedTestCosts(data, BAL = FALSE, Dosanjh3TB = FALSE,
  return = "matrix")
}
\arguments{
\item{data}{IDEA data set (perhaps bootstrap sampled or imputed)}

\item{return}{Return a list of matrix.}
}
\value{
Total test cost statistics
}
\description{
Uses the pathway frequencies for the NICE idealised pathway (chest X-ray, culture, smear, clinical judgement)
and whether the observed Dosanjh final diagnosis is what would be expected from the idealised pathway.
Assume that three sputum samples are taken so at least 3 culture and smear tests.
Inclusion of BAL would further include this procedure cost and the additional culture and smear tests.
Where it is not (usually when the 4 measurement are insufficient to discriminate) then we use
the raw data and just sum the cost of the observed tests. The subjectivity come in when deciding
exactly what the expected diagnosis is. In most cases it is clear.
We could remove patients who do not have all 3 tests so there may be some bias introduced,
but can use an imputed completed sample (prefered).
}
\details{
This is the function used in \code{\link{boot.calcIdealisedTestCosts}}.
}

