% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealised-pathway-time-to-diag.R
\name{calcIdealisedTimetoDiag}
\alias{calcIdealisedTimetoDiag}
\title{Calculate NICE Idealised Time to Diagnosis}
\usage{
calcIdealisedTimetoDiag(data, BAL = FALSE, Dosanjh3TB = FALSE,
  use_times_in_data = FALSE, return = "list")
}
\arguments{
\item{data}{IDEA study data set (perhaps bootstrap sampled or imputed).}

\item{use_times_in_data}{Use the raw times in the IDEA data or the idealised times.}

\item{return}{Return a list of matrix.}
}
\value{
Times to diagnosis statistics
}
\description{
Uses the pathway frequencies for the NICE idealised pathway (chest X-ray, culture, smear, clinical judgement)
and whether the observed Dosanjh final diagnosis is what would be expected from the idealised pathway.
}
\details{
For the pathway derived from the data only, this is like a `double-imputation' because we impute missing test results to match-up with the NICE guidelines
and then we use imputed times to diagnosis.

For when we use predefined values we avoid some of the messiness of the time to test data. At the moment we just use some hard-coded expected values.
Future work could use some statistic from the multiple imputations, like the median or mode.
We could also (hardcode) data derived averaged instead.
}
