% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-of-tests-fns.R
\name{calcPatientCostofTests}
\alias{calcPatientCostofTests}
\title{Sum the Cost of All Specified Treatments for Each Patient}
\usage{
calcPatientCostofTests(data, COSTS, x3 = TRUE)
}
\arguments{
\item{data}{IDEA study data set}

\item{COSTS}{Named vector each test/procedure and unit costs.}

\item{x3}{Should the smear and culture be multiplied by 3? It is standard practice to perform 3 of each with a sputum sample.}
}
\value{
Single value per patient
}
\description{
\code{calcPatientCostofTests} sums the total cost of diagnostic tests in the pathway.
}
\details{
If a BAL is taken we include the cost of the proceduce and the additional cost of a culture and smear if used for these.
}

