% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{calcRiskFactorScore}
\alias{calcRiskFactorScore}
\title{Calculate a risk factor score for each patient}
\usage{
calcRiskFactorScore(data)
}
\arguments{
\item{data}{}
}
\value{
data
}
\description{
\code{calcRiskFactorScore}
}

