% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{calcSumRow}
\alias{calcSumRow}
\title{calcSumRow}
\usage{
calcSumRow(data, test, subsetvar, namesdt, val)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{calcSumRow}
}

