% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean IDEA clinical data.}
\usage{
cleanData(data)
}
\arguments{
\item{data}{Individual patient records}
}
\value{
data
}
\description{
\code{cleanData} cleans IDEA clinical data for analysis
}

