% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{combineTestResults}
\alias{combineTestResults}
\title{Combine test results}
\usage{
combineTestResults(test1, test2)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{combineTestResults} returns joined tables of inputs and predicted outcomes
}

