% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-COSTdistns.R
\docType{data}
\name{COST.distns}
\alias{COST.distns}
\title{Probability Distribution Parameter Values for the Test and Procedure Costs}
\format{A list with 15 test/procudure names:
\describe{
  \item{TBcult}
  \item{Smear}
  \item{IGRA}
  \item{TSTres}
  \item{TSPOT}
  \item{QFN}
  \item{CXR}
  \item{CSF}
  \item{BAL}
  \item{HistBiop}
  \item{NeedleAsp}
  \item{PCR}
  \item{CT}
  \item{MRI}
  \item{PET}
}}
\source{
\url{}
}
\usage{
COST.distns
}
\description{
The distribution can be characterised as \code{gamma} for a gamma distribution,
\code{unif} for a uniform distribution or \code{none} for a single point value.
}
\keyword{datasets}
