% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{create.indivpathmatrix}
\alias{create.indivpathmatrix}
\title{create.indivpathmatrix}
\usage{
create.indivpathmatrix(data)
}
\arguments{
\item{data}{}
}
\value{
pathdata
}
\description{
\code{create.indivpathmatrix}
}

