% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{createCombinedGrid}
\alias{createCombinedGrid}
\title{createCombinedGrid}
\usage{

  createCombinedGrid(dir = "C:/Users/Nathan/Dropbox/TB/IDEA/output_data/idealpathway_grids/")
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{createCombinedGrid}
}

