% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{createDecisionTree}
\alias{createDecisionTree}
\title{createDecisionTree}
\usage{
createDecisionTree(data, test.DT, diag_obs.DT, diag_est.DT)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{createDecisionTree}
}

