% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crosstabDosanjh.R
\name{crosstabDosanjh}
\alias{crosstabDosanjh}
\title{Create contingency table by Dosanjh categories of data and ideal pathways}
\usage{
crosstabDosanjh(out)
}
\arguments{
\item{out}{output from \code{createDecisionTree}}
}
\value{
table
}
\description{
\code{crosstabDosanjh} returns
}

