% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decisionTree_plot.R
\name{decisionTree.plot}
\alias{decisionTree.plot}
\title{Plot data paths of decision tree with frequencies}
\usage{
decisionTree.plot(data, ysteps = "exp", out = NA, newplot = TRUE,
  col = "grey")
}
\arguments{
\item{data}{casted frequency along each pathway}

\item{ysteps}{which branching pattern}
}
\value{
none
}
\description{
\code{decisionTree.plot}
}

