% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decisionTree_plot.R
\name{decisionTreeMultiple.plot}
\alias{decisionTreeMultiple.plot}
\title{Plot multiple data paths of decision tree with frequencies}
\usage{
decisionTreeMultiple.plot(dt, names, outcol, ncols = 4)
}
\arguments{
\item{dt}{casted frequency along each pathway}

\item{names}{which branching pattern}

\item{outcol}{}

\item{ncols}{to use 4 columns or 2}
}
\value{
none
}
\description{
\code{decisionTreeMultiple.plot}
}

