% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgrids.R
\docType{data}
\name{diag_Sxest}
\alias{diag_Sxest}
\title{Idealised pathway input-diagnosis/action look-up table}
\format{...}
\source{
\url{}
}
\usage{
diag_Sxest
}
\description{
Idealised pathway input-diagnosis/action look-up table
}
\keyword{datasets}
