% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{EstimateSymptomsResolved}
\alias{EstimateSymptomsResolved}
\title{Estimate Symptoms Resolved by Whether There Was a Response to Treatment or Self-Cured}
\usage{
EstimateSymptomsResolved(data, symptomsEndDates.names, symptoms.names,
  drugReviewPeriod)
}
\arguments{
\item{data}{IDEA study data set}

\item{symptoms.names}{E.g. cough, weight loss, etc}

\item{drugReviewPeriod}{Time to follow-up by clinicians}
}
\value{
data
}
\description{
Checks if there is a response to treatment or self-cured.
Uses \code{all} as conservative estimates to check if the time between the observed end dates and earliest time that a patient was
put on treatment is smaller than the time to follow-up by clinicians.
}

