% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{fillInEndOfTreatmentDate}
\alias{fillInEndOfTreatmentDate}
\title{Fill-in End Of Treatment Dates}
\usage{
fillInEndOfTreatmentDate(data)
}
\arguments{
\item{data}{IDEA study data set}
}
\value{
data
}
\description{
If we have the duration of treatment and start of treatment but no
end date then we can calculate the end of treatment date.
}

