% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeToDiagnosisTable.R
\name{get.allDataTestCombinations}
\alias{get.allDataTestCombinations}
\title{Get All Data Test Combinations}
\usage{
get.allDataTestCombinations(data)
}
\value{
data
}
\description{
\code{get.allDataTestCombinations}
}
