% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{get.meansdOutcomeByCovariate}
\alias{get.meansdOutcomeByCovariate}
\title{get.meansdOutcomeByCovariate}
\usage{
get.meansdOutcomeByCovariate(field, value, cutoff = 8, outcome)
}
\arguments{
\item{field}{}

\item{value}{}

\item{cutoff}{}

\item{outcome}{}
}
\value{
data
}
\description{
\code{get.meansdOutcomeByCovariate}
}

