% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeToDiagnosisTable.R
\name{get.meansdStartToDiagByCovariate}
\alias{get.meansdStartToDiagByCovariate}
\title{get.meansdStartToDiagByCovariate}
\usage{
get.meansdStartToDiagByCovariate(field, value, cutoff = 8)
}
\value{
data
}
\description{
\code{get.meansdStartToDiagByCovariate}
}
