% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv-dectree-sampling.R
\name{get.sd.from.normalCI}
\alias{get.sd.from.normalCI}
\title{Get standard deviation from Normal confidence interval}
\usage{
get.sd.from.normalCI(n, x_bar = NA, upperCI = NA, lowerCI = NA)
}
\arguments{
\item{n}{Sample size}

\item{x_bar}{Mean}

\item{upperCI}{Upper 95% CI}

\item{lowerCI}{Lower 95% CI}
}
\value{
sd
}
\description{
Get standard deviation from Normal confidence interval
}

