% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeToDiagnosisTable.R
\name{get.statStartToDiagByCovariate}
\alias{get.statStartToDiagByCovariate}
\title{get.statStartToDiagByCovariate}
\usage{
get.statStartToDiagByCovariate(field, value, stat)
}
\value{
data
}
\description{
\code{get.statStartToDiagByCovariate}
}
