% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{get.uniqueDataTestCombinations}
\alias{get.uniqueDataTestCombinations}
\title{get.uniqueDataTestCombinations}
\usage{
get.uniqueDataTestCombinations(pathdata)
}
\value{
pathdata
}
\description{
\code{get.uniqueDataTestCombinations}
}

