% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{getDateFrequencies}
\alias{getDateFrequencies}
\title{Get the Date Frequencies}
\usage{
getDateFrequencies(colnames, data)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{getDateFrequencies} returns joined tables of inputs and predicted outcomes
}

