% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{GroupDiagOutcomes}
\alias{GroupDiagOutcomes}
\title{Group Together Diagnosis Outcomes}
\usage{
GroupDiagOutcomes(data)
}
\arguments{
\item{data}{IDEA study data set}
}
\value{
data
}
\description{
[TODO] do over-lapping groups, not mutually exclusive
}

