% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEAdectree_simple.R
\name{IDEAdectree.simple}
\alias{IDEAdectree.simple}
\title{IDEA Decision Tree Calculation}
\usage{
IDEAdectree.simple(data, nsim = 1000,
  costDistns = COST.distns.allerror, prevalence = 0.25, cutoff = 1,
  FNtime = 42, FNdist = TRUE, SENS = 0.9, SPEC = 0.9,
  SENSvar = 0.005, SPECvar = 0.005, c.ruleout = 100,
  name.ruleout = NA, quant = 0.5, utility = NA, N = nrow(data),
  wholecohortstats = FALSE)
}
\arguments{
\item{data}{IDEA study data}

\item{nsim}{Number of sample points. Default: 1000}

\item{costDistns}{List of distribution names and parameter values for each test/procedure}

\item{prevalence}{As a probability. Default: 0.25}

\item{cutoff}{Clinical judgement threshold}

\item{FNtime}{False negative follow-up time}

\item{FNdist}{Should false negative time to follow-up distribution be used (logical)}

\item{SENS}{Sensitivity Rule-out test}

\item{SPEC}{Specificity Rule-out test}

\item{SENSvar}{Sensitivity variance}

\item{SPECvar}{Specificiaty variance}

\item{c.ruleout}{Rule-out test unit cost}

\item{name.ruleout}{Name of rule-out test to get at distribution}

\item{quant}{Quantile value of time to diagnosis and costs}

\item{utility}{QALY adjustment utility due to active TB}

\item{N}{Number of patients. The number in the data is used as default.}

\item{wholecohortstats}{Should the output stats be the total or per patient}
}
\value{
Health and cost realisations
}
\description{
This is a parred down version of the full model that is used in the paper.
This is specifically for the HTA chapter.
It probabilistically incorporates sampling variability (and standard pathway cost and time to diagnosis).
}
\examples{

library(bcea)
dat1 <- IDEAdectree.simple(data, cutoff = 0.4, specificity = 0.8)
dat2 <- IDEAdectree.simple(data, cutoff = 0.4, specificity = 0.9)
dat3 <- IDEAdectree.simple(data, cutoff = 0.4, specificity = 0.99)

dat$e <- cbind(dat1$e, dat2$e[,2], dat3$e[,2])
dat$c <- cbind(dat1$c, dat2$c[,2], dat3$c[,2])

intlabels <- c("Current", "Enhanced specificity=0.8", "Enhanced specificity=0.9", "Enhanced specificity=0.99")

m <- bcea(e=dat$e, c=dat$c, ref=1, interventions = intlabels)
contour2(m, wtp=WTP, graph = "ggplot2", ICER.size=2, pos=c(0.9,0.1), xlim=c(-5,20), ylim=c(-100,500)) + ggtitle("")
summary(m)
}
