% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPI.R
\name{INMB}
\alias{INMB}
\title{Incremental Net Monetary Benefit for IDEA enhanced diagnostic pathway}
\usage{
INMB(spec = 0.9, sens = 0.9, Ctest = 100, pathcost = 838, A = 30000,
  q = 0.67, prev, FNtime, prop_highrisk)
}
\arguments{
\item{Ctest}{Unit cost of rule-out test}

\item{pathcost}{Routine in-practice total cost (time and tests) for a non-TB patient (433+11*55*0.67)}

\item{A}{Cost-effectiveness threshold (per year)}

\item{q}{QALY}

\item{prev}{Cohort prevalence. An 'uncertain' parameter}

\item{FNtime}{An 'uncertain' parameter (days)}

\item{prop_highrisk}{Threshold high risk predictive probability. An 'uncertain' parameter}
}
\value{
INMB
}
\description{
This a kind of super slimmed-down version of \code{make.ruleoutTable.pre()}.
}

