% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{is.diagByClinicalFeatures}
\alias{is.diagByClinicalFeatures}
\title{Distiguish Between When Clinical Features Used to Rule-in or Rule-out TB}
\usage{
is.diagByClinicalFeatures(data)
}
\arguments{
\item{data}{}
}
\value{
data
}
\description{
Add a new column with \code{TB, other,} or \code{NA}.
}
