% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{joinLevels}
\alias{joinLevels}
\title{Read from database extract.}
\usage{
joinLevels(codes, lookuplist)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
updated full database
}
\description{
\code{joinLevels} takes Access database and updates one of the tables
}

