% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{joinWithLookups}
\alias{joinWithLookups}
\title{join data set with demographic look-up tables.}
\usage{
joinWithLookups(data)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
updated full database
}
\description{
\code{joinWithLookups} takes Access database and updates one of the tables
}

