% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ruleoutTable-bysubgroups.R
\name{make.ruleoutTable.bootsample}
\alias{make.ruleoutTable.bootsample}
\title{Make rule-out table of bootstrapped samples of patients}
\usage{
make.ruleoutTable.bootsample(nboot = 2, Ctest)
}
\arguments{
\item{Ctest}{}
}
\value{
out
}
\description{
Iterate over bootstrap samples and row-bind results (potentially a high number of rows).
}
