% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ruleoutTable-fixed.R
\name{make.ruleoutTable.fixed}
\alias{make.ruleoutTable.fixed}
\title{make.ruleoutTable.fixed DEPRECATED}
\usage{
make.ruleoutTable.fixed(thresh = seq(from = 1, to = 0.7, by = -0.01),
  Ctest = c(10, 50), FNcost = 0, FNtime = 42L, ruleouttime = 1L,
  pathreturn = 1L, qaly = 0.67, A = 55, npatients = nrow(data),
  cat4percent = table(data$DosanjhGrouped)[4] * 100/npatients,
  comb = NA, cat3TB = TRUE, cat4propfollowup = 0,
  prop_highrisk = 0.72, stat = "Median", model = "testFirst")
}
\arguments{
\item{thresh}{test sensitivity and specificity}

\item{Ctest}{cost of rule-out test}

\item{FNcost}{false negative cost}

\item{FNtime}{false negative time}

\item{ruleouttime}{time for rule-ou test result}

\item{pathreturn}{does a ruled-out individual return to the pathway?}

\item{qaly}{QALY for disease}

\item{A}{cost of one day in full health}

\item{npatients}{number of patients in cohort}

\item{cat4percent}{percent of patients in Dosanjh category 4}

\item{comb}{combined sensitivity, specificity and rule-out test cost array}

\item{cat3TB}{are Dosanjh category 3 patients all active TB?}

\item{cat4propfollowup}{proportion of negative Dosanjh category 4 patients, not immediately on standard pathway, who are followed-up at 6 weeks (alpha). We assume that all active TB cases are certain to be followed-up.}

\item{prop_highrisk}{proportion of negative rule-out test patients put immediately on standard pathway (gamma)}

\item{stat}{Which statistic to use for time and cost estimate (e.g. Mean, Median, 1st Qu.)}

\item{model}{Which model structure/tree design to use
(\code{highriskAtStart} is remove randomly selected proportion before rule-out test,
\code{testFirst} is test everyone first)}
}
\value{
list
}
\description{
\code{make.ruleoutTable.fixed} creates the times and costs of a diagnositic pathway for
suspected active TB with and without an initial rule-out test, split by Dosanjh category.
}
\seealso{
\code{\link{make.ruleoutTable.pre}}
}
