% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ruleoutTable-bysubgroups.R
\name{make.ruleoutTable.FNtime}
\alias{make.ruleoutTable.FNtime}
\title{Make rule-out table for multiple follow-up times}
\usage{
make.ruleoutTable.FNtime(FNtime.seq = c(7, seq(0, 60, by = 1)), Ctest)
}
\arguments{
\item{Ctest}{}
}
\value{
out
}
\description{
Iterate over a range of follow-up times and row-bind results (potentially a high number of rows).
}

