% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_ruleoutTable-pre.R
\name{make.ruleoutTable.pre}
\alias{make.ruleoutTable.pre}
\title{make.ruleoutTable.pre}
\usage{
make.ruleoutTable.pre(thresh = seq(from = 1, to = 0.5, by = -0.01),
  Ctest = c(1, 100, 200, 300, 400, 500, 600), FNcost = 0, FNtime = 42,
  ruleouttime = 1L, pathreturn = 1L, qaly = 0.67, A = 55,
  npatients = nrow(data), cat4percent = table(data$DosanjhGrouped)[4] *
  100/npatients, comb = NA, cat3TB = TRUE, cat4propfollowup = 0,
  prop_highrisk = 0.4, stat = "Median", model = "pretest.fixed")
}
\arguments{
\item{thresh}{test sensitivity and specificity}

\item{Ctest}{cost of rule-out test}

\item{FNcost}{false negative cost of true diagnosis or cost to start of standard pathway}

\item{FNtime}{false negative time to true diagnosis or start of pathway}

\item{ruleouttime}{time taken for rule-ou test result}

\item{pathreturn}{does a ruled-out individual return to the pathway? false positives return to standard pathway Y=1/N=0}

\item{qaly}{QALY for disease}

\item{A}{cost of one day in full health}

\item{npatients}{number of patients in cohort}

\item{cat4percent}{percent of patients in Dosanjh category 4}

\item{comb}{combined sensitivity, specificity and rule-out test cost array}

\item{cat3TB}{are Dosanjh category 3 patients all active TB?}

\item{cat4propfollowup}{proportion of negative Dosanjh category 4 patients, not immediately on standard pathway, who are followed-up at 6 weeks (alpha). We assume that all active TB cases are certain to be followed-up.}

\item{prop_highrisk}{Minimum predictive probability of TB risk score (delta)}

\item{stat}{Which statistic to use for time and cost estimate (e.g. Mean, Median, 1st Qu.)}

\item{model}{Which model structure/tree design to use (
\code{pretest.fixed} is include highest risk proportion with clinical judgement before rule-out test (fixed proportion risk factor independent),
\code{pretest.var} is include highest risk proportion with clinical judgement before rule-out test (risk factor dependent),
\code{posttest.fixed} is test everyone first and include randomly selected proportion then include highest risk proportions,
\code{posttest.var} is test everyone first then remove highest risk proportion,
\code{pretest.var.sensspec.var} is remove highest risk proportion before rule-out test and modify sensitivity and specificity proportions wrt subset case-mix.)}

\item{[previously,}{proportion of patients put on standard pathway by clinical judgment (gamma)],}
}
\value{
list
}
\description{
\code{make.ruleoutTable.pre} creates the times and costs of a diagnositic pathway for
suspected active TB with and without an initial rule-out test, split by Dosanjh category.
}

