% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{make.testPerformanceMetricsTable}
\alias{make.testPerformanceMetricsTable}
\title{Make test performance metrics table}
\usage{
make.testPerformanceMetricsTable(data, prop_highriskVECTOR = c(0.2, 0.4,
  0.6), specificityVECTOR = c(0.9, 0.95, 0.99), sens = 0.9)
}
\arguments{
\item{prop_highriskVECTOR}{Clinical judgement threshold as high risk patient}

\item{specificityVECTOR}{Rule-out test specificity}
}
\value{
matrix
}
\description{
Includes true and false negatives and positive and negative predictive values.
}
