% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeFollowupPlots.R
\name{makeFollowupPlots}
\alias{makeFollowupPlots}
\title{makeFollowupPlots}
\usage{

  makeFollowupPlots(data = "../../output_data/TBdata_prepd_CLEANED_190315.RData")
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{makeFollowupPlots} returns joined tables of inputs and predicted outcomes
}

