% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC-CEplane.R
\name{maketable.CEAC_estimates}
\alias{maketable.CEAC_estimates}
\title{Print a Table of Cost-Effectiveness Acceptability Curve Estimates}
\usage{
maketable.CEAC_estimates(dat.subset, cost, sens)
}
\arguments{
\item{sens}{}
}
\value{
none
}
\description{
Print a Table of Cost-Effectiveness Acceptability Curve Estimates
}

