% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv-dectree-sampling.R
\name{MoM.gamma}
\alias{MoM.gamma}
\title{Method of Moments Gamma Distribution Parameter Transformation}
\usage{
MoM.gamma(mean, var)
}
\arguments{
\item{mean}{Mean}

\item{var}{Variance}
}
\value{
shape, scale
}
\description{
Method of Moments Gamma Distribution Parameter Transformation
}
\seealso{
MoM.beta
}
