% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPI.R
\name{plot.EVPI}
\alias{plot.EVPI}
\title{Plot Expected Value of Information curves}
\usage{
\method{plot}{EVPI}(EVPIsens)
}
\arguments{
\item{EVPIsens}{output from \code{calc.EVPI}}
}
\description{
Plot Expected Value of Information curves
}
\seealso{
\code{\link{calc.EVPI}}
}
