% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-followup-fn.R
\name{plot.FollowUpChart}
\alias{plot.FollowUpChart}
\title{plot.FollowUpChart}
\usage{
\method{plot}{FollowUpChart}(data, col = 1, lty = 1, lwd = 2, xlim,
  xlab = "Time (days)", ylab = "", order = "bydrugstart", legend = TRUE,
  dropNA = TRUE, axisID = TRUE, ...)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{plot.FollowUpChart} returns joined tables of inputs and predicted outcomes
}

