% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_riskprofile.R
\name{plot.riskprofile}
\alias{plot.riskprofile}
\title{Plot Risk Profile Discrete Distribution}
\usage{
\method{plot}{riskprofile}(data, prop_highrisk, FNtime, title = NA)
}
\arguments{
\item{data}{IDEA study data set.}

\item{prop_highrisk}{Threshold of clinical judgment.}

\item{FNtime}{Follow-up delay for false negative TB cases.}

\item{title}{}
}
\value{
none
}
\description{
Cost and probability of final outcomes (along each distinct path) in decision tree.
}

