% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC-CEplane.R
\name{plotCEAC.xSens.groupSpecCost}
\alias{plotCEAC.xSens.groupSpecCost}
\title{Plot Cost-effectiveness Acceptability Curves}
\usage{
plotCEAC.xSens.groupSpecCost(dat.temp)
}
\arguments{
\item{dat.temp}{}
}
\description{
x-axis of rule-out test sensitivity and grouped
by specificity and cost of test
}
