% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC-CEplane.R
\name{plotCEplane}
\alias{plotCEplane}
\title{Plot Cost-Effectiveness Plane With Contours}
\usage{
plotCEplane(data, N = 1000, Ctest, thresh, table = FALSE)
}
\arguments{
\item{data}{IDEA study dataset}

\item{N}{Number of random draws sample points}

\item{Ctest}{Unit cost of rule-out test}

\item{thresh}{Sensitivity and specificity of rule-out test (numeric)}

\item{table}{}
}
\value{
none
}
\description{
Plot Cost-Effectiveness Plane With Contours
}
\examples{

data("TBdata_clinical_cleaned")
plotCEplane(data, N=1000, 400, 0.8)
}

