% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepDataForDecTree.R
\name{prepDataForDecTree}
\alias{prepDataForDecTree}
\title{prepDataForDecTree}
\usage{
prepDataForDecTree(data, data.names)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{prepDataForDecTree} returns joined tables of inputs and predicted outcomes
}

