% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwayComparison.R
\name{pwayComparison}
\alias{pwayComparison}
\title{Compare data and idealised care pathways.}
\usage{
pwayComparison(pdata,
  dir = "C:/Users/Nathan/Dropbox/TB/IDEA/output_data/from IDEAdectree")
}
\arguments{
\item{data}{Individual patient records}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{pwayComparison} returns joined tables of inputs and predicted outcomes
}

