% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwayComparison.R
\name{pwayComparison.simple}
\alias{pwayComparison.simple}
\title{A simplified version of comparison of care pathways.}
\usage{
pwayComparison.simple(data, test.DT, form = "Dosanjh")
}
\arguments{
\item{data}{Individual patient records}

\item{test.DT}{(idealised) decision tree object returned by \code{createDTobj()}}

\item{form}{formula identifier between grouping by final Dosanjh category (default), observed presumptive treatment decision (step1Diag) or pooled (all)}
}
\value{
list length 2
}
\description{
\code{pwayComparison.simple} returns one pathway aggregated table and median time to diagnoses
}

