% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{relabelDecTreeOutcomes}
\alias{relabelDecTreeOutcomes}
\title{relabelDecTreeOutcomes}
\usage{
relabelDecTreeOutcomes(dt, join_names)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{relabelDecTreeOutcomes} returns joined tables of inputs and predicted outcomes
}

