% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{rm.TooEarly}
\alias{rm.TooEarly}
\title{Remove Dates That are Before the Earliest Time For the Given Suspected TB Episode}
\usage{
rm.TooEarly(data, Date.names, maxtime)
}
\arguments{
\item{maxtime}{}
}
\value{
dates
}
\description{
Remove Dates That are Before the Earliest Time For the Given Suspected TB Episode
}
\seealso{
\code{\link{calc.Date.min}}
}
