% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{rmPatientsInCleaning}
\alias{rmPatientsInCleaning}
\title{Remove Patients Records In Data Cleaning}
\usage{
rmPatientsInCleaning(data)
}
\arguments{
\item{data}{IDEA study data set}
}
\value{
data
}
\description{
Due to duplication, before 2013-08-31 (last date of complete records), not excluded, not EPTB only,
patients not treated before date of first test
}
\details{
##TODO##
should we remove patients whos `means of diagnosis' was IGRA?
}
