% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-of-tests-fns.R
\name{sample.distributions}
\alias{sample.distributions}
\title{Sample from Standard Distributions}
\usage{
sample.distributions(param.distns)
}
\arguments{
\item{param.distns}{List of distribution names and their respective parameter values}
}
\value{
vector of sample points
}
\description{
Supply a list of defined distributions and one sample realisation is taken of each.
}

