% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensspec_env.R
\docType{data}
\name{senspec.env}
\alias{senspec.env}
\title{An Environment Containing the Sensitivity and Specificity Means and Variances}
\format{An object of class \code{environment} of length 48.}
\usage{
senspec.env
}
\description{
Single value variable
}
\keyword{datasets}
