% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgrids.R
\docType{data}
\name{testsdummy.allIdeaTests}
\alias{testsdummy.allIdeaTests}
\title{Idealised pathway input-diagnosis/action look-up table}
\format{...}
\source{
\url{}
}
\usage{
testsdummy.allIdeaTests
}
\description{
Idealised pathway input-diagnosis/action look-up table
}
\keyword{datasets}

