% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBALdata.R
\name{tidyCSFdata}
\alias{tidyCSFdata}
\title{Combine CSF information}
\usage{
tidyCSFdata(data)
}
\arguments{
\item{data}{}
}
\value{
data
}
\description{
The CSF data is poorly recorded/structured and so
create separate smear and culture CSF fields in to 2 types BAL/non-BAL.
}
\details{
Searches on keywords "CSF" in TypeSmear, Bodysite_smear, Othertest1Name and Othertest2Name,
and TBcultsite, Bodysite_culture, Othertest1Name and Othertest2Name.
Used for extra-pulmonary TB only.

TODO
}
