% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{writeAllFreqTables}
\alias{writeAllFreqTables}
\title{writeAllFreqTables}
\usage{
writeAllFreqTables(out, dir)
}
\arguments{
\item{out}{joined individual data and idealised pathway}

\item{dir}{output folder}
}
\value{
list of frequency tables
}
\description{
\code{writeAllFreqTables}
}

