% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{addLevel_Nottaken}
\alias{addLevel_Nottaken}
\title{Prepend Not Taken as Extra Level}
\usage{
addLevel_Nottaken(data, names)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
data Updated full database
}
\description{
\code{joinWithLookups} takes Access database and updates one of the tables.
}
