% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{as.Date.POSIX}
\alias{as.Date.POSIX}
\title{Convert Different Classes to Date Class}
\usage{
\method{as.Date}{POSIX}(date)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
data Updated full database
}
\description{
\code{as.Date.POSIX} takes Access database and updates one of the tables.
}

