% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{assert.dateinpast}
\alias{assert.dateinpast}
\title{Assert dates are in the past}
\usage{
assert.dateinpast(data, testDate.names)
}
\arguments{
\item{data}{}

\item{testDate.names}{}
}
\description{
\code{assert.dateinpast} Checks whether all of the dates all in the past
}

