% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{calc.statOutcomeByCovariate}
\alias{calc.statOutcomeByCovariate}
\title{calc.statOutcomeByCovariate}
\usage{
calc.statOutcomeByCovariate(field, value, stat, outcome)
}
\value{
data
}
\description{
\code{calc.statOutcomeByCovariate}
}
