% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPatientCostofTests.R
\name{calcPatientCostofTests}
\alias{calcPatientCostofTests}
\title{Sum the cost of all specified treatments for each patient}
\usage{
calcPatientCostofTests(testdata)
}
\arguments{
\item{testdata}{dataset test results}
}
\value{
single value per patient
}
\description{
\code{calcPatientCostofTests} takes Access database and updates one of the tables
}

