% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{calcRiskFactorScore}
\alias{calcRiskFactorScore}
\title{Calculate a risk factor score for each patient}
\usage{
calcRiskFactorScore(data)
}
\value{
data with additional column
}
\description{
using a logistic regression calculate model probability predictions.
}
\details{
##TODO##
add clinician or hospital history covariate, maybe multilevel?
}
