% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean IDEA Study Clinical Data Wrapper}
\usage{
cleanData(data)
}
\arguments{
\item{data}{Individual patient records from IDEA}
}
\value{
data
}
\description{
\code{cleanData} is a high-level function to clean IDEA clinical data for analysis.
There are rather a lot of assumptions and imposed contraints on values.
So is a bit subjective.
}
