% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{create.indivpathmatrix}
\alias{create.indivpathmatrix}
\title{create.indivpathmatrix}
\usage{
create.indivpathmatrix(data)
}
\value{
pathdata
}
\description{
\code{create.indivpathmatrix} create a table of each distinct pathway through the decision tree
}
