% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{createDTobj}
\alias{createDTobj}
\title{Create Decision Tree object}
\usage{
createDTobj(rules, input)
}
\arguments{
\item{rules}{a matrix which is the decision tree rules so combinations of results and an associated outcome}

\item{input}{a string of which columns names are the branches i.e. tests}
}
\value{
A list of matrices of branch input cominations, outcomes (decisions) and the combined data
}
\description{
\code{createDTobj}
}

