% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTimeToDiagnosis.R
\name{estimateTimeToDiagnosis}
\alias{estimateTimeToDiagnosis}
\title{Rule-based Time to Diagnosis Estimation}
\usage{
estimateTimeToDiagnosis(data)
}
\arguments{
\item{data}{Individual patient records}
}
\value{
data with the \code{start.to.diag} column appended
}
\description{
\code{estimateTimeToDiagnosis} estimates time to diagnosis for each patient using simple pre-defined rules.
This time is not directly available in the data but can be got at in most cases indirectly.
}
\details{
The rules for estimating time to diagnosis are in priority order:
\enumerate{
 \item {If culture +ve then \code{min}(report time, start treatment) [previously, take report time only.]}
 \item {What is means of diagnosis? Use corresponding times. Latest date of those for the means of diagnosis used.}
 \item If diagnosis date is too late (130 days after first test) then use \code{DateDiagCon} if <130 days
 \item If diagnosis date still \code{NA} then use  \code{DateDiagCon}.
 \item {could use \code{min(DateDiagCon, lasttestDate)} isntead?}
 }
}
