% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateTimeToDiagnosis.R
\name{estimateTimeToDiagnosis}
\alias{estimateTimeToDiagnosis}
\title{Rule-based time to diagnosis estimation}
\usage{
estimateTimeToDiagnosis(data)
}
\arguments{
\item{data}{Individual patient records}
}
\value{
timeToDiag
}
\description{
\code{estimateTimeToDiagnosis} returns estimates of time to diagnosis for each patient
}

