% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{fillInEndOfTreatmentDate}
\alias{fillInEndOfTreatmentDate}
\title{Fill-in End Of Treatment Dates}
\usage{
fillInEndOfTreatmentDate(data)
}
\arguments{
\item{data}{}
}
\value{
data
}
\description{
\code{fillInEndOfTreatmentDate}
}

