% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DecisionTreeIDEA-fn.R
\name{freqTable.DecisionTree}
\alias{freqTable.DecisionTree}
\title{freqTable.DecisionTree}
\usage{
freqTable.DecisionTree(decntree, subsetvar)
}
\arguments{
\item{decntree}{}

\item{subsetvar}{}
}
\value{
one combined frequency table of each decision outcomes
}
\description{
\code{freqTable.DecisionTree}
}

