% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{get.meansdOutcomeByCovariate}
\alias{get.meansdOutcomeByCovariate}
\title{get.meansdOutcomeByCovariate}
\usage{
get.meansdOutcomeByCovariate(field, value, cutoff = 8, outcome)
}
\value{
data
}
\description{
\code{get.meansdOutcomeByCovariate}
}
