% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{get.uniqueDataTestCombinations}
\alias{get.uniqueDataTestCombinations}
\title{get.uniqueDataTestCombinations}
\usage{
get.uniqueDataTestCombinations(pathdata)
}
\arguments{
\item{pathdata}{}
}
\value{
pathdata
}
\description{
\code{get.uniqueDataTestCombinations}
}

