% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeTable.R
\name{join.stratifiedOutcomeTables}
\alias{join.stratifiedOutcomeTables}
\title{join.stratifiedOutcomeTables}
\usage{
join.stratifiedOutcomeTables(data, outcome)
}
\value{
res.all
}
\description{
\code{join.stratifiedOutcomeTables}
}
