% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{joinLevels}
\alias{joinLevels}
\title{Read From Database Extract}
\usage{
joinLevels(codes, lookuplist)
}
\arguments{
\item{Patients}{Single table to update the data total database}
}
\value{
codes Updated full database
}
\description{
\code{joinLevels} takes Access database and updates one of the tables.
}

