% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{joinWithLookups}
\alias{joinWithLookups}
\title{Join Data Set With Demographic Look-up Tables}
\usage{
joinWithLookups(data)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
data Updated full database
}
\description{
\code{joinWithLookups} takes Access database and updates one of the tables.
}

