% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFromDatabaseExtract.R
\name{readFromDatabaseExtract}
\alias{readFromDatabaseExtract}
\title{Read from database extract.}
\usage{
readFromDatabaseExtract(Patients)
}
\arguments{
\item{Patients}{single table to update the data total database}
}
\value{
updated full database
}
\description{
\code{readFromDatabaseExtract} takes Access database and updates one of the tables
}

