% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromDatabaseExtract.R
\name{readFromDatabaseExtract}
\alias{readFromDatabaseExtract}
\title{Read From original IDEA Database Extract and Flatten (Tidy)}
\usage{
readFromDatabaseExtract()
}
\value{
updated full data set
}
\description{
\code{readFromDatabaseExtract} takes Access database and updates one of the tables
to give a combined full data set with defined list of fields included.
}
\details{
The main Access extract (in Excel tabs) is always the same original one but the Patient table (tab in Excel) has been updated
and so the latest version of this is joined to the other data.
Part of the reason for doing it like this is because don't know how to convert a whole Access database to Excel,
only a sheet at a time.
##TODO##
}
