% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data-fns.R
\name{rmPatientsInCleaning}
\alias{rmPatientsInCleaning}
\title{Remove Patients Records In Data Cleaning}
\usage{
rmPatientsInCleaning(data)
}
\arguments{
\item{data}{}
}
\value{
data
}
\description{
\code{rmPatientsInCleaning}
}

