% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{summaryTable}
\usage{
summaryTable(datat, dir = NULL)
}
\arguments{
\item{dir}{filename to save output}

\item{data}{Individual patient records}
}
\value{
dataframe
}
\description{
\code{summaryTable} gives basic statistics for patient groupings in IDEA data extract
}

