% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{summary Table}
\usage{
summaryTable(datat, dir = NULL)
}
\arguments{
\item{data}{Individual patient records}

\item{allgrids}{The idealised pathway look-up tables}
}
\value{
The joined tables with aggregated outcome wrt Dosanjh categories
}
\description{
\code{summaryTable}
}

